backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "httpx==0.27.2",
    "openai==1.51.2",
    "oss2==2.19.1",
    "psycopg2-binary==2.9.10",
    "pypdf==5.1.0",
    "pytest>=9.0.3",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "sqlalchemy==2.0.36",
    "uvicorn[standard]==0.30.6",
]
